y-2022/src/test/java/io/github/aj8gh/aoc/y22/d3/Day3Test.java
package io.github.aj8gh.aoc.y22.d3;

import static org.junit.jupiter.api.Assertions.assertEquals;

import io.github.aj8gh.aoc.util.InputProvider;
import java.util.List;
import java.util.stream.Stream;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

class Day3Test extends InputProvider {
  @ParameterizedTest
  @MethodSource(value = INPUT_PROVIDER_PART_1)
  void part1(List<String> input, int expected) {
    var day3 = new Day3();
    var actual = day3.part1(input);
    assertEquals(expected, actual);
  }

  @ParameterizedTest
  @MethodSource(value = INPUT_PROVIDER_PART_2)
  void part2(List<String> input, int expected) {
   var day3 = new Day3();
    var actual = day3.part2(input);
    assertEquals(expected, actual);
  }

  private static Stream<Arguments> inputProviderPart1() {
    return Stream.of(
        Arguments.of(reader().getExample(DAY_3).asStringList(), 157),
        Arguments.of(reader().getInput(DAY_3).asStringList(), 7990)
    );
  }

  private static Stream<Arguments> inputProviderPart2() {
    return Stream.of(
        Arguments.of(reader().getExample(DAY_3).asStringList(), 70),
        Arguments.of(reader().getInput(DAY_3).asStringList(), 2602)
    );
  }
}
